RecvJob::Execute 메서드에서 recv 함수의 반환 타입을 ssize_t로 변경 CMakePresets.json에서 Visual Studio 설명 제거 LICENSE 파일 삭제 CMakeLists.txt 파일 추가

+++ src/lib-core/src/ServerJob.cpp
@@ -158,8 +158,8 @@ bool RecvJob::Execute()
             return false;
         }
 
-        Int32 result = recv(mSession->GetSocketFD(), mRecvBuffer.GetRearBufferPtr(),
-                            mRecvBuffer.GetDirectWriteSize(), 0);
+        ssize_t result = recv(mSession->GetSocketFD(), mRecvBuffer.GetRearBufferPtr(),
+                              mRecvBuffer.GetDirectWriteSize(), 0);
         if (result == 0) {
             mSession->SetReceiveState(false);
             mSession->Disconnect();
